pyproject.toml
[project]
name = "weaveflow"
version = "0.1.0a1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "graphviz>=0.21",
    "networkx>=3.5",
    "pandas>=2.3.1",
    "pip>=25.2",
    "pygraphviz>=1.14",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "ruff>=0.12.9",
    "tomli>=2.2.1",
]

[tool.pytest.ini_options]
markers = [
    "smoke: marks tests as smoke tests (quick subset of tests).",
    "weavegraph: marks tests related to the WeaveGraph functionality.",
    "legend: marks tests for the graph legend.",
    "edges: marks tests for graph edge creation.",
    "empty: marks tests for handling empty tasks.",
]

# --- DEFINE GLOBAL LINTING SETTINGS ---
[tool.ruff]
line-length = 91
target-version = "py313"
exclude = [
    ".venv",            # Excluded by default but be explicit
    "__pycache__",      # Same
    ".ruff_cache",      # Same
    ".pytest_cache",    # Same
    "./assets",
] 

# --- DEFINE FORMATTER SETTINGS ---
[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

# --- DEFINE LINTER SETTINGS ---
[tool.ruff.lint]
select = [
    # Core Pyflakes and pycodestyle rules
    "F", "E", "W",
    # Import sorting (isort)
    "I",
    # Modernize your code (pyupgrade)
    "UP",
    # Bug-finding (flake8-bugbear)
    "B",
    # Code simplification (flake8-simplify)
    "SIM",
    # Naming conventions (pep8-naming)
    "N",
    # Docstring conventions (pydocstyle) - CHOOSE ONE SECTION
    # "D", # Uncomment if you want to enforce docstrings
    # Code complexity (mccabe)
    "C90",
    # Catch print statements (flake8-print)
    "T20",
    # Security lints (flake8-bandit)
    "S",
    # General best practices and error-proneness (pylint)
    "PLC", "PLE", "PLR", "PLW",
    # Pytest style
    "PT",
    # Ruff-specific rules
    "RUF",
    # Pandas-specific rules
    "PD",
    # Use Path instead of os,
    "PTH",
]

ignore = [
    "S101",
    "N803",
    "PLR0913",
    "PLR2004",
    "UP038",
    "RUF007",
    "PD901",
    "PLW2901",
    "S108",
]